Add Bman test to user_example and tidy up the output.
This also adds a calm_down() step after testing so that FQRNI messages can
be flushed before drivers exit, otherwise a subsequent run will fail to
load because of an unclean portal. This is normally not an issue because
portal drivers are persistent and FQRNI messages are ignored, so nothing
waits for them.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>