Fix assignment of channel ids (due to UIO ordering).
Our current dts results in cpus 0..7 getting portals 8,1,2,...,7. For the most
part, this "just works" because the environment variables mapping cpus to UIO
devices was adjusted accordingly. However when the driver is initialising a
portal, it also needs to know the id for its portal's dedicated channel, which
is currently computed from the cpu number using hard-coded assumptions that
broke when the dts changed. This fails in the case of portal 0, causing tests
running on cpu 0 to schedule FQs not to themselves, but to the portal running in
the kernel!

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>