Make barriers+spinlocks non-polling (use pthread.h)
The sync logic uses pthread_cond_*(), and spinlocks use pthread_mutex_*(),
avoiding polling on atomic variables in both cases.

This makes for less noisy threading, but may slightly increase latencies of
locking and synchronisation. The assumption is that the latter are not fast-path
considerations and that overheads from busy-waits are more of a concern.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>