Add dpasync.sh for syncing with kernel src.
NB, this was used against the 2.1 branch to pull in know bug fixes (and things
related to errata workarounds) as well as push out things fixed here into 2.1.
However, we are not adding such fixes/updates to the kernel code used in usd for
now, so running this with a usl/bsp.git checkout will generate lots of noise.
Please ignore for now.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>