poc: rework 'tx' linkage
The per-interface structure contains its Rx FQs and its Tx FQ. Prior
to this commit, the Rx FQ was initialised with a pointer to the Tx FQ,
whereas we now point Rx back to the interface structure. This is to
free up possibilities for using multiple Tx FQs, but it means an extra
level of indirection, in theory (though the explicit prefetch is
arguably an improvement on the previous code if ever accessing Tx was
a cache-miss).

The reason for making this change prior to implementing multiple TX
FQs is to allow this change to be tested for performance regression on
its own.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>