poc: support 1-to-1 Rx/Tx FQ mapping.
Enabled by default. Embeds the Tx FQ into the Rx FQ's state. This
means 32 Tx FQs as well. Note that turning on HOLDACTIVE with this
option would provide end-to-end order preservation per-flow.

This can't be combined with the "one FQ per cpu per interface"
configuration. It's believed this new mapping should supercede it
anyway, which will make the code simpler.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>