poc: allow for fewer Tx than Rx FQs.
This retains the PERRX configuration, where Rx and Tx FQ *objects* are
coupled, however it allows the Tx FQIDs to be collapsed, ie. so that
multiple Tx FQ objects may be backed by the same FQID. The default app
settings are now to use, per-interface, 32 Rx FQs and 16 Tx FQs.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>