apps: cleanup sync logic, code-reduction.
The apps/common.[ch] code no longer needs the notion of a "master" thread. The
sync_all() usage in 'user_example' is moved to that app and out of common.[ch].
The global initialisation in 'poc' is mostly performed prior to thread-startup
so that the one-time initialisation in thread-startup (which can't be performed
in advance because it requires access to a portal) doesn't need synchronisation.
Nonetheless, a simple barrier is used in thread startup just to avoid degenerate
behaviour on the simulator (h/w is unaffected).

This also bypasses FQ-allocator initialisation in 'poc' and MAC-enablement in
'user_example'.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>