poc: implement crude CLI with stats.
You can enter "q", "foobar", or "dump" from the cmd-line, and only the
latter is useful. The dump is currently an unfiltered dump of all app
state. For each cpu, it dumps per-cpu state for each of the
interfaces, after which it dumps an accumulated set of stats across
all cpus. The inner per-interface dumps show all "2drop" stats on a
single line followed by "2fwd" stats for each of the RX_HASH_SIZE FQs,
one-per-line.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>