qman: do not set EQCR_ITR in run-to-completion mode.
This only affects attempts to enqueue when EQCR is full. Linux may put the
thread to sleep, so the EQCR_ITR-generated interrupt is needed to ensure
EQCR_CI is rechecked. On run-to-completion models (where fast-path is not
interrupt-driven), the enqueue will be retried in a qman_poll() loop, so
the EQCR_ITR writes are needless and will trigger excess interrupts.

(In user-space, this is just a case of unnecessary writes. On LWE, it will
probably cause spurious interrupts in loaded scenarios.)

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>