qman: "productise" adaptive EQCR::CI throttling.
If CONFIG_FSL_QMAN_ADAPTIVE_EQCR_THROTTLE is defined, the low-level code
tracks CI updates in a histogram and a new function, qm_eqcr_cce_avg_x10(),
allows the high-level to query the average CI update whilst resetting the
histogram. The high-level code implements an adaptive throttle on CI
updates using this functionality.

All of the above is compiled out by default, as it does not yet provide a
benefit except in cases where s/w outperforms Fman egress processing, and
even then the benefit is limited. (Whereas performance when s/w is the
bottleneck is degraded.)

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>