Reorganise FQ-allocator initialisation.
The __fqalloc_init() interface is driver internal, so move it out of
fsl_qman.h. However this means calling it from within the driver too,
whereas driver init is strictly thread-local. Instead, add a USD-specific
wrapper to fsl_usd.h, and call it from apps/common.c - and achieve the
latter using a barrier during thread setup (because portals need to be
initialised prior to the fq-allocator).

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>