Remove old build system before adding new one.
This is primarily to avoid git patches that make the new build system
look like a modification of the old one. The diff of such change would
be every bit as ugly as it would be pointless.

This also removes the dangling qman and bman "Makefile"s that were
never being used (left-overs from the driver port from the kernel).

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>