Simplify the build system.
ARCH stuff (Makefile.powerpc) and the build magic (Makefile.rules) are
integrated into the single top-level Makefile.

A lot of duplication is reduced by some reorganisation. The order is
also clearer now.

g++/flex/bison support was removed to facilitate the simplifications,
though it should now be straightforward to add support back in if the
need arises.

Comments have been removed now that the logic is clearer, and the
README.BuildSystem file now documents the inner-workings so that the
Makefile remains compact.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>