Add build dependencies on Makefile(.sub) files.
Dependencies for all compilable objects now also include;
  * the Makefile.sub that defines them,
  * the top-level Makefile.
In this way, any changes that would affect *how* the objects are to be
built will trigger a rebuild.

Also, when (re)constructing a library, remove it first. This covers the
corner case where a Makefile is edited to remove an object file from a
library. Note however that the lib doesn't need to be declared as
dependent on the Makefile, because the other objects are already and so a
Makefile change will trigger their recompilation and thus the
reconstruction of the lib).

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>