Rename [fsl_]shmem to dma_mem
We should have a DMA-able memory allocator and the rest of the code shouldn't care how the memory is
obtained/allocated from the system. As of right now we're still using the fsl_shmem driver but
ultimately it will be replaced by HugeTLBFs. Upon such a transition the code using DMA-able memory
should not need to change in any way

Signed-off-by: Emil Medve <Emilian.Medve@Freescale.com>