reflector: optimise the worker thread messaging.
In trying to debug some performance degradation (that unexpectedly
became a study of different toolchain versions), I was twiddling with
the alignment and structure of the "struct worker" type, as well as the
"struct worker_msg" type that is within it. The latter was previously a
sub-struct of the former, but slightly improved performance is seen if
it is a distinct allocation and dereferenced instead. Also, both structs
need to be cache-aligned (to avoid cache-thrashing), and the pointer
needs to be the first element of the worker struct (presumably the
compiler finds a neat optimisation if it has no offset to take into
account).

Note that this apparently trivial fiddling generates observable results
because the worker threads are all calling check_msg() in their critical
loop, which needs to be low-overhead in the likely path (where there are
no messages to process - messages only occur in reaction to CLI
commands);

  while (check_msg(worker)) {
      qman_poll();
      bman_poll();
  }

After toolchain distinctions were removed, the performance impact of
this change is about half a percent (of throughput in the single-cpu
case).

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>