reflector: improve pool-channel selection algorithm.
The use of an LFSR provides some easy "randomness" in the way hashed rx
FQIDs are assigned to the available pool-channels, in order to avoid
"thundering herds". However it makes no guarantees that they'll be even
spread! This change makes sure that for each sequence of n rx FQs (where
n is the number of pool channels), each pool channel occurs is assigned
exactly once.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>