fman: generalise the MAC initialisation.
Rather than a single API (__mac_enable_all()) to parse the device-trees,
map the registers, and enable the MACs, the parsing and mapping now
happens at initialisation time, which builds up a linked-list of structs
representing the mapped interfaces. The macs can therefore be
efficiently turned on and off at run-time by iterating the list.

reflector is also modified for the API change. Additionally, the
"macs_on" and "macs_off" commands are added to the CLI to turn MACs on
and off at run-time.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>