reflector: add depletion-based flow-control
When the primary buffer pool enters depletion, the fman MACs are
disabled, and are reenabled when the pool exits depletion. If the system
is seriously oversubscribed, this may thrash quite quickly so the
"freq_xtcl <n>" CLI command allows the logging frequency to be reduced
(eg. to 5, 1000, ...). The default value is 1.

The RFL_DEBUG_DEPLETION_FLOW symbol controls this behaviour but it is
off by default.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>