reflector: turn on bpool-depletion-based flow-control.
This facilitates testing of ZLT, because depletion itself seems to
trigger the most degenerate behaviour (rather than the use of PFDRs per
se, as was previously assumed). It seems possible that egress-processing
gets starved in such circumstances, but only further testing will tell.

In the mean time, defaulting the pool-depletion flow-control mechanism
to on means that as soon as traffic is below the ZLT, loss-less
performance resumes quickly (despite there being no evidence that use of
PFDRs has stopped).

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>