build: example environment for building outside ltib.
It is easier and quicker to develop (and spot warnings) when compiling
usdpaa directly, as compared to building it from ltib. However,
dependencies on ltib-installed packages require some extra guidance,
because the usdpaa tree itself doesn't know where these external things are
unless it is told. When ltib is building usdpaa, this information is fed
in, but we have to do it ourselves if building in isolation.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>