reflector: compile out CGR support by default.
For the release, there is no flow-control support, so the only purpose
for compiling CGR support is to track the Rx and Tx fill-levels in the
system. However, these incur a slight cost (both latency and
scalability) in Qman h/w, because each enqueue and dequeue operation
involves a locked update to the relevant CGR. So for obtaining the best
performance, we disable this by default.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>