fman: clarify the "fman" driver interface
There is some confusion between;
a) the fman driver, which discovers the Fman network interfaces (and their
   MACs) via the device-tree and allows them to be configured, and
b) the non-XML part of usdpa_syscfg, which discovers the Fman network
   interfaces via the device-tree, and exposes information about them.

The usdpa_syscfg layer extracts a number of parameters that the fman driver
doesn't, and vice versa. There is certainly no good reason to parse these
device-tree nodes twice, independently. The first step in organising it is
to clarify the Fman driver interface, which deals with "fman_if" (if ==
interfaces) objects. The next step will be to additionally extract all the
params that usdpa_syscfg needs within the Fman driver's walk of the device
tree, and have usdpa_syscfg pick those up from the Fman driver, rather than
parsing the device-tree a second time. This will in turn clarify
usdpa_syscfg, which will combine XML info (from the fmc_netcfg layer) and
device-tree info (from the fman driver) into the unified config, exposed
to apps.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>