fman: allow individual i/faces to be manipulated.
The driver-internal list of interfaces is (partially) exposed. For now
only the mac_type parameter is visible, because it's the only one the fman
driver parses so far, as it's required to determine the offset to the
register for enabling/disabling Rx and Tx. The CCSR mapping and other
driver internals are not visible from the API.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>