reflector: integrate with the config layer.
A few issues had to be addressed for this, so it's a fairly big change.

One major issue is that there has, for some time, been a latent bug that
was disabling context-stashing when we thought all along that it was
enabled - the macro to compute the number of cachelines of a struct was
rounding down to 0, so when that struct became cacheline-aligned the
macro produced the right answer: 1. The problem is that context-stashing
is decreasing performance! This needs investigation, but for now I'm
disabling context-stashing properly to get best numbers for the release.

The method for allocation and using Tx FQs has changed, as it is the
only hard-coded input that the config layer doesn't specify (because any
Tx FQIDs will do). The nature of the forwarding logic is (a little) more
closely modelled on a "generic network processing application", see the
comments in the struct definitions for more detail.

The two paths required by the config layer for XML files are compiled
into the app via defines in the Makefile.am. These can be overriden by
environment variable if required, but the app finds the installed config
files by default.

Due to time constraints, known issues;

* RFL_CGR support hasn't been tested, it may not even compile. It is
  disabled by default.

* "quit" at the CLI whilst traffic is flowing can hose the portals,
  requiring a reboot. If traffic is stopped, the app can exit and
  restart.

* A cpu can be "add"ed and "rm"ed from the CLI, but it can't be "add"ed
  a second time. The app must be "quit" and relaunched before that cpu
  can be used again. (Need to clean up the UIO "fd", which requires some
  missing hooks.)

* Some paths in the code are missing error-handling.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>