dma_mem: fix bug in allocator metadata
There was a typo in the linked-list logic for the 'dma_mem' allocator
that had failed to crash anything, so far, by a mix of good fortune and
a lack of application support for repeated setup and teardown of
threads.

In addition to fixing this bug, the extra debugging that was added to
solve this problem is also committed - define DEBUG_DMA_ALLOC to compile
this in.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>