reflector: reap dead threads from the CLI
Previously if a thread exited, the CLI portion of the code would not
notice, so the user would have to "rm" it. This patch causes the CLI
loop to scan the running threads each time it is about to display the
prompt, to check for dead threads. This is similar to shells like bash,
where if a background task has exited, the shell will only notice once
the user has hit "enter".

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>