qbman: get rid of 16K-alignment requirement.
The driver was using bitwise-OR of the cache-enabled base-address and the
required offset to access cache-enabled portal registers. Switching this
to addition removes the requirement for the mappings of cache-enabled
regions to match the h/w alignment, however we still require that rings
be sufficiently aligned to perform carry-based wrapping. (Maximum is for
DQRR, where at least 2K-alignment is required.) The minimum mapping
guarantees are already page-based (4K) so this is fine.

Also, avoid the use of in_be32()/out_be32() for access 32-bit register
fields - on linux (at least), these accessors imply heavyweight memory
sync()s. Use a volatile dereference instead - the code already includes
any required syncs.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>