mmap: minimise hard-coded mmap() addresses.
The Qman/Bman portal mappings are now completely chosen by the kernel,
so they pose no risk of a hard-coded address conflicting with the
run-time state of the application's VMA. However, the dma_mem mapping must
continue to use MAP_FIXED because the TLB1 hack requires that the mapping
be size-aligned. Nonetheless, the DMA interface now uses a run-time
variable for the mapped base-address rather than a compile-time #define,
so any performance impact associated with a kernel-chosen mapping (eg. in
vtop/ptov conversion) will be observable immediately. (The removal of this
final hard-coding will probably require HugeTLB support.)

Note, the change to use a real variable for the dma_mem base-address
(rather than a #define) causes some header-dependency issues. Eg. the
dma types and base addresses are required within compat.h, which is
usually the lowest-level header (everything else is dependent on it). As
such, all DMA-related definitions are now moved to dma_mem.h and compat.h
includes it implicitly.

As part of the header cleanup, the legacy dependency of compat.h on other
headers (fsl_bman.h, fsl_qman.h, fsl_usd.h) is now removed, requiring some
minor adjustments to the "#include"s in some other files.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>