reflector: initial split into ppac/ppam.
PPAC = Packet-Processing Application Core
PPAM = Packet-Processing Application Module (eg. reflector)

The current split is a first-step. The core application code is separated out
into a *header* that is inlined by the application module. Eg. reflector
declares the structs and hooks ("ppam_*") that PPAC needs, then includes
"ppac_main.c" to instantiate the core application code. It is done this way
because, in theory, this should yield no performance loss relative to the
existing reflector app prior to the split. If so, we can then look at more
elegant extensions of this split and quantify the degree of any performance
overheads.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>