build: improve dependencies on Makefiles.
We were declaring each compilable object as dependent on both the
Makefile.am that defined it as well as the top-level Makefile. This
ignores intermediate Makefiles when the directory structure is deeper
than 1-ply. The fact that we now use SUBDIRS to control which Makefiles
are in or out makes this worse, because changes to a Makefile potentially
impact sub-directories too. (Previously this wasn't the case because we
were "find"ing Makefile.am files without a SUBDIRS-style interface.)

This patch changes the dependencies for compilable objects to no longer
be explicitly dependent on the top-level Makefile, just the local
Makefile.am. Instead, Makefiles are themselves declared as being dependent
on the parent Makefile that included them via SUBDIRs, causing a
dependency chain all the way to the top-level.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>