build: use a top-level Makefile.am
Now that the build is SUBDIRS-driven, it is clearer to use a top-level
Makefile.am to declare "apps" and "drivers" as SUBDIRS, rather than to
have a special notion of top-level DIRS built into the build system
itself.

For example, this would also allow for READMEs or other files in the
top-level directory to be declared as installable objects in the same way
that is done in any other directory.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>