dma_mem: de-linuxify the implementation, comment.
This code originated from something I'd previously written in a kernel
context, and it was possible to re-use it due to our USDPAA compat.h stubs
for kmalloc, GFP_KERNEL, etc. However this is USDPAA-specific code, so it
makes more sense to have it declared in terms of native/standard/posix
user-space interfaces.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>