qman: fix "park" support, and suppress bogus CRITs
When DQRR is in discrete-consumption mode, there is *no* CI tracking in
the driver, because there is no logical equivalent to a "CI" index when
you are potentially completing ring entries out-of-order. In the
qman_high.c logic, the parking of "CI" entry is really supposed to park
the *current* entry - ie. the entry that the internal "cursor" pointer is
tracking. This renames the low-level driver primitive accordingly.

Also, when tearing down a portal (eg. when 'reflector' stops a thread),
bogus "CRIT:Ignoring completed DQRR entries" messages would be logged.
These are due to qm_dqrr_finish() verifying that CI is up-to-date, without
checking whether we are actually tracking CI at all! This changes that
check to be modal.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>