build: don't build sec+simple crypto by default.
While churn in definitions, headers, [etc] is on-going, it is preferable
to leave this code (as with ipfwd) out of the default build, because it is
maintained by other people and so necessarily will lag behind as it gets
updated to match underlying changes. Please locally uncomment the relevant
lines in the Makefiles to re-enable these build targets.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>