qman: fix mgmt-command structure padding.
Commit a30786d0682e31635f2ae0df6754ddb9905461d8 made the "qm_fqd" struct
slightly bigger, but without reducing the corresponding padding in the MC
command and response structure types. As a result, copies by the driver
to/from the portal command/response regions would read or write unused
padding data 2 bytes beyond the designated cacheline boundary. Due to the
layout of the cache-enabled region of the portal, this is in fact harmless
and went undetected when testing on h/w, but the simics model asserts a
DML exception for these illegal accesses.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>