sanity: add 'check_types' utility.
This is a quick and dirty struct-checker that I wrote to catch the DML
exception observed on simics (due to commit a30786d0). NB, "sizeof()" is
not adequate for these tests, because sizeof() calculates how much space
a struct would need if declared as an array, which may imply padding due
to alignment requirements of some structure elements. Instead, a wrapper
struct is declared with a single byte "suffix" and the "__packed"
attribute, and the offset to the suffix is used to determine the
structure size.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>