dma_mem: use newer /dev/fsl-usdpaa-shmem device.
This change is inter-dependent on the corresponding kernel change
("fsl_usdpaa_shmem: rename driver, and rework it").

* There are no more DMA_MEM_* symbols in the exported conf.h, they are all
  in the internal conf.h now (ie. there's no more API impact).
* As the DMA region is no longer a hard-coded constant, the physical
  base-address and size are obtained via ioctl().

Similarly, the applications must seed buffer-pools themselves. A commit
for ppac (reflector) will follow that drains the buffer pools of any stale
contents and then seeds them according to the available DMA range and a
default set of buffer requirements.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>