ppac: seed buffer pools from the application.
The <internal/conf.h> defines the BPIDs, buffer sizes, and buffer counts
to use for USDPAA apps (these match the settings that used to be in the
device-tree and seeded statically from within the kernel). This change
makes the ppac app perform the buffer-pool seeding on start-up. It also
drains the pool of any stale values prior to doing so, to improve
restartability. This is also more robust if the kernel network driver ever
adds its own buffers to those pools during start up (or requires that the
device-tree provide some static allocations), so long as it is purely for
boot-up/errata-workaround purposes.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>