ppac: allow PPAM to override the CLI prompt.
Weakly-links the PPAC-supplied default CLI prompt, so that a PPAM can
override it by declaring its own string of the same name. Reflector uses
this to set the prompt to "reflector> ", matching the prompt that existed
prior to the recent ppac changes. (For one thing, this is easier to manage
for automated testing - the simple ">" prompt own its own is too fiddly to
distinguish.)

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>