ppac: reduce the IRQ blocking to 1/100s.
After a worker experiences a period of inactivity, it goes into IRQ mode
and sleeps. The sleep has a timeout simply to ensure that inter-thread
messaging is serviced regularly. This was initially set to 1 second, as it
provides visible clues as to its operation - tearing down 7 idle threads
stutters noticably, given that each teardown pauses until the sleeping
thread wakes up, whereas tearing down 7 busy threads is nearly
instantaneous because they're never sleeping. This functionality is
working, so reducing the timeout to 1/100s eliminates the visible clue and
makes the system more responsive to control from the CLI. Given that a
busy thread can handle up to 9 million packets a second, checking for
messages in between sleeps of 1/100s should not be significantly more
power-hungry than sleeping in blocks of 1s, and certainly not when
compared to busy processing.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>