qbman/ppac: reduce success-path verbosity.
When portals are successfully setup and torn down, we get one line of
output for the portal plus another for the IRQ hookup, neither of which is
very useful. (In particular, when creating and deleting threads from the
CLI, this is very noisy.) Instead, rely on the existing "Thread alive"
message printed by PPAC when threads are created, and add a corresponding
message for when threads are deleted.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>