qman: isolate 32-bit-specific conversions.
The p5020 kernel port of the Qman driver already supports a FQ look-up
table for 64-bit CPUs (it's needed because we frequently only have 32-bits
of passthrough state available in the h/w programming interface). Until
the kernel and USDPAA drivers are resynchronised, this change isolates the
casting into two inlines. These casts squash some warnings from the
build-checkers until the eventual driver resync removes the need to do
this.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>