build: correct accumulation of BINS and LIBS
In processing Makefile.am files, the build will call pre_process_target()
for all dist_DATA entries, process_bin() for all bin_PROGRAMS entries, and
process_lib() for all lib_LIBRARIES entries. So far so good.
pre_process_target() adds each installable file to TO_INSTALL, but
process_bin() and process_lib() don't do the same for bins and libs,
instead the directory-level function, process_dir(), does it. The result
is that declaring installable files in the top-level Makefile.am works
fine, but declaring libs or bins does not (it only works in lower-level
directories), so this corrects that.

NB, this is not a big deal, there's not much need to declare lib or bin
targets in the top-level directory, however I noticed this issue when I
was trying to add an ad-hoc test program to the build+install, and if a
customer was to try to link their own targets this issue could show up
there too, so it's better to just fix it.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>