ppac: add -n/--non-interactive argument
With this option, ppac apps (reflector) won't insist on controlling stdin
for the CLI, which allows the app to be backgrounded without it stopping.

Eventually it would make sense to support another option to read initial
commands from a file, so that it could be used in combination with "-n".

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>