ppac: allow multiple threads per CPU.
Each thread is given a distinct unique identifier ('UID') rather than
using the CPU as a unique idenitifer, thus allowing multiple threads to be
identified. The log messages are updated to display UID as well as CPU,
and the "rm" CLI command can look-up by UID (using "uid:...") as well as
the existing numeric CPU lookup. When removing by CPU(s), when a named CPU
has more than one thread, the first such thread is removed.

Also, testing the failure case revealed that all barrier-based messaging
was unstable - if the thread exits before responding to a message, the app
hangs, and this is the case when failing to initialise a portal. So the
messaging is now polling based, and handles reaping of dead threads.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>