of: single parse from C, no scripts.
The script-based trick was cute, but was slow, and was also failing on
corner-cases on new SoC device-trees. For one thing, phandles were being
passed as ASCII characters to scripts for comparing against binary files -
which only worked so long as the phandles stayed in a cooperative range.

This instead parses the procfs device-tree once on startup from C code
only. An internal representation of the tree is made in a tree-like data
structure, with an additional linear/flat list of the directory entries to
help with lookups, like of_find_compatible_*(). Also, when building the
tree, the directory nodes cache lookups to "interesting" properties
("linux,phandle", "compatible", [etc]), making it quicker for related
APIs.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>