qman: fix EQCR_CI cleanup corner case.
The optimised enqueue consumption tracking model causes a CI prefetch to
occur when producing an enqueue to the second-to-last EQCR ring entry, and
then the CI is read (and then invalidated) when producing an enqueue to
the last EQCR entry. If we destroy the portal in the corner case where a
prefetch was issued but there was no subsequent enqueue to trigger the
follow-up processing, then the last-issued enqueue will remain unconsumed
and trigger the usual warnings when tearing down the portal. Ie. the one
update call in the tear-down logic processes a stale CI prefetch and
invalidates it, but doesn't issue nor process a fresh read. Calling the
update twice eliminates this clean-up problem.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>