of: fix a teardown corruption.
Support for teardown of the "of" driver layer was not been used, so the
fact it was broken went unnoticed. Due to other bug fixes that are to
follow, this needs to be rectified.

The top-level "dir" node in the device-tree representation is a static
global variable so it shouldn't be free()'d. This change ensures that the
recursive logic to tear down the tree destroys directory nodes "of
children, by parents", thus avoiding any harm to the root node.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>