Init/finish the "of" driver directly from apps.
Previously the device-tree interface ("of") was internal-only, and the
driver was only initialised via the application library
"usdpaa_netcfg_layer". Due to a design error in the USDPAA/UIO port of the
Qman and Bman drivers, they will soon be dependent on device-tree handling
as well, and so it no longer makes sense to make device-tree usage
dependent on a specific application-config layer (especially with respect
to the order in which the layers are initialised).

So there is now an external <usdpaa/of.h> header as well as the internal
one, but only the of_init() and of_finish() APIs are moved to it. It is
the responsibility of all apps that will use device-tree-dependent layers
to call of_init() prior to using any of those layers, and to call
of_finish() after tearing them down (if leak-free cleanup is desired).
Every main() function in the source tree (except "sanity/check_types")
has been updated for this, and so the hook within usdpaa_netcfg_layer.c
has been removed.

Also;
* some protections are added to the "of" code via a new macro, WARN_ON(),
  that produces a one-time only warning for each instance where it is
  declared. Any failure to initialise the device-tree layer before it is
  used will generate error log messages, and this will also catch any
  double initialisations or teardowns.
* the newly-exported of_init() API now takes no parameters but instead
  defers to of_init_path() with a default path ("/proc/device-tree"), for
  improved convenience. This path can be overriden either by calling
  of_init_path() instead of of_init(), or by redefining
  OF_INIT_DEFAULT_PATH prior to including the <usdpaa/of.h> header.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>