qbman: initialise via device-tree
This changes the way portals are discovered and managed, and in particular
changes the expected naming of UIO-based Qman/Bman portal devices. As
such, this patch is mutually inter-dependent with the linux kernel patch
that renames the UIO devices in the same way;
"qbman: rename UIO devices by cell-index"

The Qman/Bman drivers now rely on the "of" driver being initialised so
that portals affine to the required CPU can be found, all previously-
missing attributes can be extracted (dedicated portal channel ID, pool
mask, cell-index), and from there the UIO device path can be derived too.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>