pme loopback app: Fix direct mode command, add extra command
The direct mode scanning command didn't have the pme ctx flag to
indicate direct mode scanning, hence flow based scanning was
happening.

Added prep_scan_2 command which is a different version of pre_scan
command. This version allows the user to specify the pattern to
send. The rest of the SUI is filled with '.' characters or the sui
size is larger than the pattern. If the sui size is smaller the pattern
is truncated.

Signed-off-by: Jeffrey Ladouceur <Jeffrey.Ladouceur@freescale.com>