qbman: default irq_sources should be zero.
On portal startup, certain irq_sources were being set as non-zero, which is
wrong because unless the app specifies otherwise, it is only expected to
processing polling-based portal duties (meaning the duties set in
irq_sources would not get processed at all). PPAC support for switching to
interrupt mode when idle was hiding this issue, because the first entry to
(and exit from) interrupt mode would clobber the faulty startup settings.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>