build: Initial integration of Klocwork static analyzer
All the KW specific foo is in a separate Makefile.kw that gets included into the top Makefile. If
Makefile.kw is not present (as it would be the case in a released package/tarball) the top Makefile
will not fail and continue graceously

To run a build with KW analysis, use the kwall target. It assumes you have the kw-user package
installed with its bin directory in your path and that is minimaly configured to acquire a license.
The analysis is incremental and the kwall target can be used during day-to-day development as it
also builds the code using the regular cross-compiler

Here is a samle installation command line:

.../kw-user-installer.9.1.0.4735.linux.sh --agree --install-dir ~/local/klocwork --klocwork-server klocwork.freescale.net:80 --license-server luge-a:5562

Signed-off-by: Emil Medve <Emilian.Medve@Freescale.com>