compat: improve portability
These are some changes that are required to support non-p4080 SoCs.
(Driver changes are to follow that depend on these.) In particular,
support for non-corenet, non-64B-cacheline systems is made possible by a
refactoring of cacheline manipulations. Also, __raw_[read|write][bl]()
accessors are added to provide compatibility with kernel changes that are
to be ported to the USDPAA drivers.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>